ml-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
scikit-learn==1.4.1
pandas==2.2.1
pymongo==4.6.2
numpy==1.26.4
spacy==3.7.4
dateparser==1.2.0
python-dotenv==1.0.1
